Move Fuzz tests into their own file
This is required for https://github.com/google/oss-fuzz/pull/11453
